Merge changes from main repository

+++ org.envirocar.app/src/org/envirocar/app/handler/PreferencesHandler.java
@@ -94,13 +94,13 @@ public class PreferencesHandler implements PreferenceConstants {
         PreferenceManager.getDefaultSharedPreferences(context).edit().putInt(PREF_LOCAL_TRACK_COUNT ,count).apply();
     }
 
-    public static Long getTotalDistanceTravelledOfUser(Context context){
-        return PreferenceManager.getDefaultSharedPreferences(context)
-                .getLong(PREF_TOTAL_DIST_TRAVELLED, 0);
+    public static Double getTotalDistanceTravelledOfUser(Context context){
+        return Double.valueOf(PreferenceManager.getDefaultSharedPreferences(context)
+                .getFloat(PREF_TOTAL_DIST_TRAVELLED, 0));
     }
 
-    public static void setTotalDistanceTravelledOfUser(Context context, Long distance){
-        PreferenceManager.getDefaultSharedPreferences(context).edit().putLong(PREF_TOTAL_DIST_TRAVELLED, distance).apply();
+    public static void setTotalDistanceTravelledOfUser(Context context, Double distance){
+        PreferenceManager.getDefaultSharedPreferences(context).edit().putFloat(PREF_TOTAL_DIST_TRAVELLED, distance.floatValue()).apply();
     }
 
     public static String getTotalTime(Context context){
